+++ app/x2mail/v/current/app/libraries/X2Mail/Engine/Actions/Messages.php
@@ -542,33 +542,6 @@ trait Messages
 			} catch (\Throwable $e) {
 				$this->logException($e);
 			}
-/*
-			if (!$oMessage->sPlain && !$oMessage->sHtml && !$oMessage->pgpEncrypted && !$oMessage->smimeEncrypted) {
-				$aAttachments = $oMessage->Attachments ?: [];
-				foreach ($aAttachments as $oAttachment) {
-//					\in_array($oAttachment->ContentType(), ['application/vnd.ms-tnef', 'application/ms-tnef'])
-					if ('winmail.dat' === \strtolower($oAttachment->FileName())) {
-						$sData = $this->ImapClient()->FetchMessagePart(
-							$oMessage->Uid,
-							$oAttachment->PartID()
-						);
-						$oTNEF = new \TNEFDecoder\TNEFAttachment;
-						$oTNEF->decodeTnef($sData);
-						foreach ($oTNEF->getFiles() as $oFile) {
-							if (\in_array($oFile->type, ['application/rtf', 'text/rtf'])) {
-								$rtf = new \X2Mail\Engine\Rtf\Document($oFile->content);
-								$oMessage->setHtml($rtf->toHTML());
-							} else {
-								// List as attachment?
-								$oMapiAttachment = new \X2Mail\Mail\Client\Attachment($sFolder, $iUid, BodyStructure);
-								$oMessage->Attachments->append($oMapiAttachment);
-							}
-						}
-						break;
-					}
-				}
-			}
-*/
 		}
 		catch (\Throwable $oException)
 		{