+++ app/x2mail/v/current/app/libraries/X2Mail/Engine/http/request/socket.php
@@ -151,7 +151,10 @@ class Socket extends \X2Mail\Engine\HTTP\Request
 		# Read body
 		$body = '';
 		if (\is_resource($this->stream)) {
-			while (!\feof($sock)) {
+			// Cap streamed bytes too (security review S4).
+			$max_bytes = $this->max_response_kb * 1024;
+			$streamed = 0;
+			while (!\feof($sock) && (!$max_bytes || $streamed < $max_bytes)) {
 				if ($chunked) {
 					$chunk = \hexdec(\trim(\fgets($sock, 8)));
 					if (!$chunk) { break; }
@@ -159,10 +162,13 @@ class Socket extends \X2Mail\Engine\HTTP\Request
 						$tmp = \fread($sock, $chunk);
 						\fwrite($this->stream, $tmp);
 						$chunk -= \strlen($tmp);
+						$streamed += \strlen($tmp);
 					}
 					"\r\n" === \fread($sock, 2);
 				} else {
-					\fwrite($this->stream, \fread($sock, 1024));
+					$tmp = \fread($sock, 1024);
+					\fwrite($this->stream, $tmp);
+					$streamed += \strlen($tmp);
 //					\stream_copy_to_stream($sock, $this->stream);
 				}
 			}