+++ app/x2mail/v/current/app/libraries/X2Mail/Engine/rtf/controlsymbol.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace X2Mail\Engine\Rtf;
-
-class ControlSymbol implements Entity
-{
-	public string $symbol;
-	public /*int|string*/ $parameter;
-}