+++ app/x2mail/v/current/app/libraries/X2Mail/Engine/rtf/controlword.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace X2Mail\Engine\Rtf;
-
-class ControlWord implements Entity
-{
-	public string $word;
-	public int $parameter;
-}