+++ app/x2mail/v/current/app/libraries/X2Mail/Engine/rtf/groups/listtable.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace X2Mail\Engine\Rtf\Groups;
-
-class ListTable extends Destination
-{
-}