+++ app/x2mail/v/current/app/libraries/X2Mail/Mail/Smtp/SmtpClient.php
@@ -177,40 +177,8 @@ class SmtpClient extends \X2Mail\Mail\Net\NetClient
 
 		try
 		{
-			$sRequest = '';
-			if (\str_starts_with($type, 'SCRAM-')) {
-				// RFC 5802 send "client-first-message" and receive "server-first-message"
-				$sRequest = $SASL->authenticate($sLogin, $sPassword, $sResult);
-				$this->logMask($sRequest);
-				$sResult = $this->sendRequestWithCheck($sRequest, 334);
-				// RFC 5802 send "client-final-message" and receive "server-final-message"
-				$sRequest = $SASL->challenge($sResult);
-				$this->logMask($sRequest);
-				$sResult = $this->sendRequestWithCheck($sRequest, 334);
-				$SASL->verify($sResult);
-				// Now end the authentication
-				$sRequest = '';
-				$this->sendRequestWithCheck($sRequest, 235);
-			} else switch ($type) {
-			// RFC 4616
-			case 'PLAIN':
-			case 'XOAUTH2':
-			case 'OAUTHBEARER':
-				$sRequest = $SASL->authenticate($sLogin, $sPassword);
-				break;
-
-			case 'LOGIN':
-				$sRequest = $SASL->authenticate($sLogin, $sPassword, $sResult);
-				$this->logMask($sRequest);
-				$sResult = $this->sendRequestWithCheck($sRequest, 334);
-				$sRequest = $SASL->challenge($sResult);
-				break;
-
-			// RFC 2195
-			case 'CRAM-MD5':
-				$sRequest = $SASL->authenticate($sLogin, $sPassword, $sResult);
-				break;
-			}
+			// OAUTHBEARER / XOAUTH2 only (NC-only SSO)
+			$sRequest = $SASL->authenticate($sLogin, $sPassword);
 			if ($sRequest) {
 				$this->logMask($sRequest);
 				$SASL->verify($this->sendRequestWithCheck($sRequest, 235));